initialization (windows)
Running 'python main.py' now adds a context menu command to file explorer in windows. A context menu command is a command that pops up in the right click selection of a file in file explorer, like '7Zip' and lets you do things with the file. Currently, it just opens up a gui window that has a placeholder button which renames the file to 'abcd' and the file extension that was there before. Both the button and the action are placeholedrs for the real gui and file transfer.