detection of all devices on network pt. 1
There is now a script that detects all devices in the network of the user. This is currently not hooked up to the rest of the application, however it successfully gets the IP and mac of other devices on the network. Later, it will also fetch the os and username.