Sender Receiver Server prototype pt. 1
There is now a (not yet finished) prototype of a sender, receiver, and server. The good thing about this model is that neither sender, nor receiver, need to know each other, just the server. Both send a uuid, and sender also sends along a file as well. If two uuids of a receiver and server match, then the file gets unlocked and sent to receiver and both sockets close. Later on, near the end of transferX, the sender will be able to right click on a file, and get a uuid on their gui. Then the receive can right click on a folder and enter uuid, and get the file transferred. The sender will also be given gui feedback of the stage of the transfer.