Sender Receiver Server prototype pt. 2
The sender server and receiver model now has many bugs fixed, and receiver correctly receives the given file from the sender. There is one uncomplete thing, which is the file is not sent with it's contents in binary correctly so it does not have the text inside of txt or the audio inside of mp3, only the name and extension.