Sender Receiver Server prototype pt. 3
The prototype model is now complete (for now) and successfully stores a UUID and file on server from a sender.py prompt, and releases it upon a receiver.py prompt containing the corresponding UUID. Server.py holds the file on a temporary folder in it's directory, and contains the full information of the file by copying it's exact bytes. Because of the 2-step process with the server, this model works with the sender and receiver being anonymous and anywhere on the world, since they do not need to know or connect to each other directly.