GUI Transfer Integration pt. 1
The TKinter GUI now works with the S/S/R Model created before. You can right click on any file, click on "TransferX", and get a UUID for your file. And on any computer globally on any network, you can right click on a folder, click on "TransferX", then enter your UUID, and receive the file on the folder.