isPermanent option
There is now a checkbox option for permanent file storage. Storing a file permanently means that even after you receive it on any computer, it does not get deleted, so you can clone it on any computer any amount of times you want, making transferX double as an online drive.